pyproject.toml
[tool.poetry]
name = "chutes-api"
version = "0.0.1"
description = "chutes.ai API/validator"
authors = ["Chutes Global Corp <dev@chutes.ai>"]
license = "MIT"
readme = "README.md"



[tool.poetry.dependencies]
python = "<3.13,>=3.9"
aiohttp = "^3.10.10"
backoff = "^2.2.1"
requests = "^2.32.3"
loguru = "^0.7.2"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
pybase64 = "^1.4.0"
orjson = "^3.10.10"
SQLAlchemy = "2.0.42"
pydantic-settings = "^2.6.0"
asyncpg = "^0.30.0"
passlib = "^1.7.4"
python-multipart = "0.0.19"
taskiq-redis = "^1.0.2"
fickling = "^0.1.3"
setuptools = "^75.3.0"
py-bip39-bindings = "^0.1.11"
argon2-cffi = "^23.1.0"
python-slugify = {extras = ["unidecode"], version = "^8.0.4"}
async-lru = "^2.0.4"
# TODO: Pin the below version
# fiber = { git = "https://github.com/rayonlabs/fiber.git" }
aiodns = "^3.2.0"
chutes = "0.3.50.rc1"
fastapi-cache2 = {extras = ["memcache"], version = "^0.2.2"}
python-socketio = {extras = ["asyncio-client"], version = "^5.11.4"}
pillow = "^11.0.0"
aioboto3 = "14.1.0"
#cryptography = "^44.0.0"
py-ed25519-zebra-bindings = "1.1.0"
jinja2 = "^3.1.4"
pyjwt = "^2.10.1"
aiocache = "^0.12.3"
ruff = "^0.9.2"
aiomcache = "^0.8.2"
async-substrate-interface = ">=1.4.0"
transformers = "<4.49.0"
dnslib = "^0.9.26"
semver = "^3.0.4"
bittensor-wallet = "^3.1.0"
cllmv = "0.0.6"
dcap-qvl = "^0.3.2"
huggingface-hub = "0.34.6"
tokenizers = "0.21.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
greenlet = "^3.1.1"
ipykernel = "^6.29.5"
typer = "^0.12.5"
pytest-asyncio = "^1.1.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint.extend-per-file-ignores]
"tests/*" = ["E402"]
"**/conftest.py" = ["E401", "E402"]

[tool.ruff]
extend-exclude = ["nv-attest"]
